admin/requirements.txt
Django==4.0
djangorestframework==3.14.0
mysqlclient==2.1.1
django-cors-headers==3.10.0
pika==1.2.0
